web/requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.3
django-environ==0.9.0
django-extensions==3.2.1
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
executing==1.2.0
flake8==6.0.0
idna==3.4
inflection==0.5.1
ipython==8.6.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.32
psycopg==3.1.4
psycopg2==2.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.13.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
tomli==2.0.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
